+++ k8s-standard/kube-loki-stack.tf
@@ -3,9 +3,17 @@ variable "loki_enabled" {
   default = false
 }
 
+variable "loki_disk_size_gb" {
+  default = 50
+}
+
+variable "loki_retention_days" {
+  default = 7
+}
+
 variable "kube_loki_stack_version" {
   type = string
-  default = "2.10.2"
+  default = "2.10.3"
 }
 
 variable "kube_loki_stack_values_override" {
@@ -31,7 +39,28 @@ resource "helm_release" "kube_loki_stack" {
   namespace = kubernetes_namespace.kube-loki-stack[0].metadata.0.name
 
   values = [
-    templatefile("${path.module}/kube-loki-stack-values.yaml", {}),
+    <<EOF
+    loki:
+      enabled: true
+
+      config:
+        table_manager:
+          retention_deletes_enabled: true
+          retention_period: ${var.loki_retention_days}d
+
+      persistence:
+        enabled: true
+        size: ${var.loki_disk_size_gb}Gi
+
+      resources:
+        requests:
+          cpu: 100m
+          memory: 750Mi
+        limits:
+          cpu: 1
+          memory: 1Gi
+    EOF
+    ,
     yamlencode(var.kube_loki_stack_values_override)
   ]
 }